+++ src/main/java/beloved/beloved/entity/OrderItem.java
@@ -14,7 +14,7 @@ public class OrderItem {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    private BigDecimal quantity;
+    private int quantity;
 
     @ManyToOne
     @JoinColumn(name = "order_id")
@@ -24,7 +24,7 @@ public class OrderItem {
     @JoinColumn(name = "product_id")
     private Product product;
 
-    public OrderItem(Long id, BigDecimal quantity, Order order, Product product) {
+    public OrderItem(Long id, int quantity, Order order, Product product) {
         this.id = id;
         this.quantity = quantity;
         this.order = order;
@@ -41,11 +41,11 @@ public class OrderItem {
         this.id = id;
     }
 
-    public BigDecimal getQuantity() {
+    public int getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(BigDecimal quantity) {
+    public void setQuantity(int quantity) {
         this.quantity = quantity;
     }
